chore: [CO-1712] test message renderer with shadow DOM (#742)
* chore: add data-testid to shadow dom wrapper

* test(shadow-dom-wrapper): add test

* test: access shadow DOM in mail preview tests

Updated mail preview tests to access the shadow DOM using the
shadowRoot property. This change ensures that the tests correctly
retrieve and verify the content within the shadow DOM of the
message renderer component.

* test(shadow-dom-wrapper): add tests for DarkReader

Added tests to verify that DarkReader is enabled when dark mode is
enabled in the ShadowDomWrapper component. The tests ensure that
the enableDarkReader function is called and that the children are
correctly rendered inside the shadow DOM.

* chore: move test into test folder

---------

Co-authored-by: Keshav Bhatt <keshavnrj@gmail.com>